% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_find_limit.R
\name{chisq_find_limit}
\alias{chisq_find_limit}
\title{Determines the hyperparameter of the THAMES for mixtures,
as described in Metodiev et al. (2025).}
\usage{
chisq_find_limit(lps, d_par)
}
\arguments{
\item{lps}{vector of unnormalized log-posterior values}

\item{d_par}{number of (free) posterior parameters}
}
\value{
the matrix param_test with one additional column
}
\description{
Determines the hyperparameter of the THAMES for mixtures,
as described in Metodiev et al. (2025).
}
\references{
Martin Metodiev, Nicholas J. Irons, Marie Perrot-Dockès,
Pierre Latouche, Adrian E. Raftery. "Easily Computed Marginal Likelihoods
for Multivariate Mixture Models Using the THAMES Estimator."
arXiv preprint arXiv:2504.21812.
}
\keyword{internal}
