% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_by_lda.R
\name{reorder_by_lda}
\alias{reorder_by_lda}
\title{Calculates the function W described in Metodiev et al. (2025).}
\usage{
reorder_by_lda(scaling, G, sims)
}
\arguments{
\item{scaling}{list of fit determined via QDA (means and covariances)}

\item{G}{number of components}

\item{sims}{n_simul x G x (u+1) array of parameters sampled from
the posterior, where
  n_simul is the number of simulations from the posterior,
  G       is the number of components,
  u       is the number of mixture component parameters
          (parameter u+1 is the mixture weight)}
}
\value{
a named list with the values of W as its element
}
\description{
Calculates the function W described in Metodiev et al. (2025).
}
\references{
Martin Metodiev, Nicholas J. Irons, Marie Perrot-Dockès,
Pierre Latouche, Adrian E. Raftery. "Easily Computed Marginal Likelihoods
for Multivariate Mixture Models Using the THAMES Estimator."
arXiv preprint arXiv:2504.21812.
}
\keyword{internal}
