% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unformly_functions.R
\name{runif_in_ellipsoid}
\alias{runif_in_ellipsoid}
\title{Uniform sampling in ellipsoid}
\usage{
runif_in_ellipsoid(n, A, r)
}
\description{
Uniform sampling on an ellipsoid or in an ellipsoid.
  The sampling \emph{in} an ellipsoid is available in arbitrary
  dimension. The sampling \emph{on} an ellipsoid is available only in
  dimension 2 or 3.
}
\keyword{internal}
