% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unformly_functions.R
\name{runif_sphere}
\alias{runif_sphere}
\alias{runif_on_sphere}
\alias{runif_in_sphere}
\title{Uniform sampling on/in sphere}
\usage{
runif_on_sphere(n, d, r = 1)

runif_in_sphere(n, d, r = 1)
}
\arguments{
\item{n}{number of simulations}

\item{d}{dimension of the space}

\item{r}{radius of the sphere}
}
\value{
The simulations in a \code{n} times \code{d} matrix.
}
\description{
Uniform sampling on a sphere or in a sphere, in arbitrary
dimension.
}
\keyword{internal}
